URGENT! Push version 1.2.1!
We missed updating this when rolling back the changes from our light entity test. As a result, any out_lvl_switch that could function as a light won’t appear in Home Assistant until this version is properly released. At that point, HA will recognize it as an update. Alternatively, a manual reinstall MAY also resolve the issue.